convert App to ts
