--- src/renderer/src/App.jsx
@@ -1,14 +1,0 @@
-import TodoDate from "./components/TodoDate.tsx";
-import ItemList from "./components/ItemList.tsx";
-import { AppStateProvider } from "./AppContext.tsx";
-
-function App() {
-  return (
-    <AppStateProvider>
-      <TodoDate />
-      <ItemList />
-    </AppStateProvider>
-  );
-}
-
-export default App;
