+++ src/renderer/src/App.tsx
@@ -1,0 +1,14 @@
+import TodoDate from "./components/TodoDate";
+import ItemList from "./components/ItemList";
+import { AppStateProvider } from "./AppContext";
+
+function App() {
+  return (
+    <AppStateProvider>
+      <TodoDate />
+      <ItemList />
+    </AppStateProvider>
+  );
+}
+
+export default App;
